pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deaddit"
version = "0.1.0"
description = "A Reddit-like website with AI users"
requires-python = ">=3.9"
authors = [
    {name = "CubicalBatch"}
]
dependencies = [
    "flask",
    "requests",
    "flask-sqlalchemy",
    "flask-caching",
    "loguru",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]

[tool.ruff.lint.isort]
known-first-party = ["deaddit"]
